refactor: organize examples by topic, add GCP/Netlify/Azure scaffolding
The flat examples/ directory had grown to 10 sibling .ts files mixing
SMTP / cache / high-level / integration use cases — hard to navigate.
Reorganize by topic so the structure mirrors the test suite and the
serverless platform layout:

  examples/
  ├── smtp/                   ← verifyMailboxSMTP usage / TLS / caching
  ├── cache/                  ← CacheStore<T> implementations (memory, Redis)
  ├── high-level/             ← verifyEmail + name / domain examples
  ├── integrations/           ← Algolia + future external integrations
  └── serverless/<platform>/  ← deploy-ready per-platform scaffolding

Renames also drop the redundant `-example` suffix (the folder name
already says "example"). Path adjustments: relative imports bumped from
`../src/...` to `../../src/...` for the moved files.

New serverless scaffolding for the platforms shipped in the previous
commit:

- examples/serverless/gcp/        — Cloud Functions 2nd gen (Functions
  Framework + gcloud deploy command)
- examples/serverless/netlify/    — netlify.toml with /api/* redirect
  + netlify/functions/validate.ts using netlifyHandler
- examples/serverless/azure/      — host.json with /api route prefix
  + src/functions/validateEmail.ts using app.http() with wildcard route

Docs updates:
- examples/README.md rewritten with the new topic groupings + serverless
  table now lists all 8 platforms
- README.md project-structure tree reflects __tests__/{unit,isolated}/,
  the new examples/<topic>/ layout, and the full adapter list
- README.md scripts section adds the new domain-specific test:smtp /
  test:cache / test:whois / test:names / test:serverless / test:cli
  filter scripts (already added in previous commit, now documented)
- README.md "Running Examples" updates paths for the new layout
