+++ examples/serverless/gcp/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "email-validator-gcp-example",
+  "version": "1.0.0",
+  "private": true,
+  "main": "index.js",
+  "scripts": {
+    "start": "functions-framework --target=validateEmail --source=.",
+    "deploy": "gcloud functions deploy validateEmail --gen2 --runtime=nodejs20 --trigger-http --allow-unauthenticated --entry-point=validateEmail"
+  },
+  "dependencies": {
+    "@emailcheck/email-validator-js": "^3.4.3",
+    "@google-cloud/functions-framework": "^3.4.0"
+  }
+}
